requirements.txt
pandas
numpy
scikit-learn
scikit-surprise
streamlit